requirements.txt
# Core
anthropic>=0.40.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0

# Web UI
streamlit>=1.30.0

# Document Parsing
pdfplumber>=0.10.0
python-docx>=1.0.0

# Web Scraping
httpx>=0.26.0
beautifulsoup4>=4.12.0

# PDF Generation
weasyprint>=60.0
jinja2>=3.1.0
markdown>=3.5.0

# Utilities
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
